+++ rotkehlchen/chain/ethereum/adex/utils.py
@@ -1,6 +1,4 @@
-from typing import Union, cast
-
-from eth_typing import HexStr
+from typing import Union
 
 from rotkehlchen.accounting.structures import Balance
 from rotkehlchen.assets.asset import EthereumToken
@@ -11,7 +9,6 @@ from rotkehlchen.serialization.deserialize import (
     deserialize_ethereum_address,
     deserialize_timestamp,
 )
-from rotkehlchen.typing import Timestamp
 
 from .typing import AdexEventDBTuple, AdexEventType, Bond, ChannelWithdraw, Unbond, UnbondRequest
 
@@ -39,8 +36,10 @@ DAI_AMOUNT_MANTISSA = FVal(10**18)
 
 # Tom pool fee rewards API constants
 TOM_POOL_FEE_REWARDS_API_URL = 'https://tom.adex.network/fee-rewards'
-PERIOD_END_AT_FORMAT = '%Y-%m-%dT%H:%M:%S.%fZ'
+TOM_POOL_FEE_REWARDS_ADX_LEGACY_CHANNEL = '0x30d87bab0ef1e7f8b4c3b894ca2beed41bbd54c481f31e5791c1e855c9dbf4ba'  # noqa: E501
+TOM_POOL_PERIOD_FORMAT = '%Y-%m-%dT%H:%M:%S.%fZ'
 
+OUTSTANDING_REWARD_THRESHOLD = FVal('0.2')
 ADEX_EVENTS_PREFIX = 'adex_events'
 
 # Defines the expected order of the events given the same timestamp and sorting
@@ -75,6 +74,7 @@ def deserialize_adex_event_from_db(
     11 - unlock_at
     12 - channel_id
     13 - token
+    14 - log_index
     """
     db_event_type = event_tuple[4]
     if db_event_type not in {str(event_type) for event_type in AdexEventType}:
@@ -82,11 +82,11 @@ def deserialize_adex_event_from_db(
             f'Failed to deserialize event type. Unknown event: {db_event_type}.',
         )
 
-    tx_hash = HexStr(event_tuple[0])
+    tx_hash = event_tuple[0]
     address = deserialize_ethereum_address(event_tuple[1])
     identity_address = deserialize_ethereum_address(event_tuple[2])
     timestamp = deserialize_timestamp(event_tuple[3])
-    pool_id = HexStr(event_tuple[5])
+    pool_id = event_tuple[5]
     amount = deserialize_asset_amount(event_tuple[6])
     usd_value = deserialize_asset_amount(event_tuple[7])
     value = Balance(amount=amount, usd_value=usd_value)
@@ -104,15 +104,15 @@ def deserialize_adex_event_from_db(
             timestamp=timestamp,
             pool_id=pool_id,
             value=value,
-            bond_id=HexStr(cast(str, event_tuple[8])),
-            nonce=cast(int, event_tuple[9]),
-            slashed_at=Timestamp(cast(int, event_tuple[10])),
+            bond_id=event_tuple[8],  # type: ignore # type already checked
+            nonce=event_tuple[9],  # type: ignore # type already checked
+            slashed_at=deserialize_timestamp(event_tuple[10]),  # type: ignore # already checked
         )
 
     if db_event_type == str(AdexEventType.UNBOND):
         if any(event_tuple[idx] is None for idx in (8,)):
             raise DeserializationError(
-                f'Failed to deserialize bond event. Unexpected data: {event_tuple}.',
+                f'Failed to deserialize unbond event. Unexpected data: {event_tuple}.',
             )
 
         return Unbond(
@@ -122,7 +122,7 @@ def deserialize_adex_event_from_db(
             timestamp=timestamp,
             pool_id=pool_id,
             value=value,
-            bond_id=HexStr(cast(str, event_tuple[8])),
+            bond_id=event_tuple[8],  # type: ignore # type already checked
         )
 
     if db_event_type == str(AdexEventType.UNBOND_REQUEST):
@@ -138,27 +138,24 @@ def deserialize_adex_event_from_db(
             timestamp=timestamp,
             pool_id=pool_id,
             value=value,
-            bond_id=HexStr(cast(str, event_tuple[8])),
-            unlock_at=Timestamp(cast(int, event_tuple[11])),
+            bond_id=event_tuple[8],  # type: ignore # type already checked
+            unlock_at=deserialize_timestamp(event_tuple[11]),  # type: ignore # already checked
         )
 
     if db_event_type == str(AdexEventType.CHANNEL_WITHDRAW):
-        # NB: `token` (event_tuple[13]) could be None, do not check below.
-        if any(event_tuple[idx] is None for idx in (12,)):
+        if any(event_tuple[idx] is None for idx in (12, 13, 14)):
             raise DeserializationError(
-                f'Failed to deserialize unbond request event. Unexpected data: {event_tuple}.',
+                f'Failed to deserialize channel withdraw event. Unexpected data: {event_tuple}.',
             )
 
-        token = None
-        if event_tuple[13] is not None:
-            try:
-                token = EthereumToken(event_tuple[13])
-            except (UnknownAsset, UnsupportedAsset) as e:
-                asset_tag = 'Unknown' if isinstance(e, UnknownAsset) else 'Unsupported'
-                raise DeserializationError(
-                    f'{asset_tag} {e.asset_name} found while processing adex event. '
-                    f'Unexpected data: {event_tuple}',
-                ) from e
+        try:
+            token = EthereumToken(event_tuple[13])   # type: ignore # type already checked
+        except (UnknownAsset, UnsupportedAsset) as e:
+            asset_tag = 'Unknown' if isinstance(e, UnknownAsset) else 'Unsupported'
+            raise DeserializationError(
+                f'{asset_tag} {e.asset_name} found while processing adex event. '
+                f'Unexpected data: {event_tuple}',
+            ) from e
 
         return ChannelWithdraw(
             tx_hash=tx_hash,
@@ -167,8 +164,9 @@ def deserialize_adex_event_from_db(
             timestamp=timestamp,
             value=value,
             pool_id=pool_id,
-            channel_id=HexStr(cast(str, event_tuple[12])),
+            channel_id=event_tuple[12],  # type: ignore # type already checked
             token=token,
+            log_index=event_tuple[14],  # type: ignore # type already checked
         )
 
     raise DeserializationError(